Kors Logic Trainer
Dark Koreth has provided a great tutorial for players to learn Logic and Signals.

Portal is located by start planet teleporters.

Within the instance players can use the consoles to access a screenshot of the Signals tab for the structure.

The TutorialsOrbit file can be used for any space structures in the future.

Note: Must press ESC to have ability to scroll and navigate other pages on the TR website.